Update some shard `u16` to `u64`
This is pretty hacky, is just serving my purpose for `BeaconChain`.
These structs will need to be fully updated to the latest spec.

+++ beacon_chain/types/src/validator_record.rs
@@ -14,7 +14,7 @@ pub enum ValidatorStatus {
 #[derive(Debug, Clone, PartialEq)]
 pub struct ValidatorRecord {
     pub pubkey: PublicKey,
-    pub withdrawal_shard: u16,
+    pub withdrawal_shard: u64,
     pub withdrawal_address: Address,
     pub randao_commitment: Hash256,
     pub randao_last_change: u64,
